remove "test" tooltip from gamewall
fix TXTcheatpath not being saved

made prompt and call you an idiot if you try to load in image that has bad dimensions.

added stuff from the cheat menu to be translatable 

maybe fixed an issue with cheatcode names being too long and running on/off out of site

un comment sam & max fix.  we were under the impression that it only worked from disc but somebody said that it is working from USB also with ios rev 10 or 13a.  i have left out the ios checks to see if it works with any ios and also it is only on if the game ID matches one of the two listed on wiiboxart.com for this game.

drink a beer