Update New Super Mario Bros. Wii patch
New Super Mario Bros. Wii checks the BCA of the disc, and if it doesn't match the expected pattern, it triggers a timer that throws a fake disc read error and logs it to the NAND (`/shared2/test2/dvderror.dat`) between 3 and 10 minutes after the game boots.

Nintendo included code that lets RVT units bypass the BCA check, so we can patch this to circumvent the check.

Thanks to blackb0x for providing the initial NTSC-U version of the patch.